Initial commit of Hashtable
This commit is the code I wrote. The CMake piece is rough but it will
compile and passes CTest... At least it does on my machine.

The CMake section is the piece whose quality I'm not thrilled with -
seems very rough to my untrained CMake eye - but this was an exercise
for myself to become familiar with C. Given thats been achieved I'm only
slightly disgruntled I wasn't able to polish the CMake piece - but I can
live with that.