Initial Commit
This commit is the foundation - multiple threads are spawned and will
print to console. If I science things to oblivion this is the reset
point.

The challenge from here is figuring out how to how to pass each
thread a reference to silverware.

Possible solutions/Notes

We could probably force it with 'unsafe' but that defeats the point.
Plus I'd be surprised if there wasn't a safe way to do this - I'm not
familiar enough with Rust.

.clone() won't work - we don't want a deep copy. To this end, we appear
to want n shallow copies where n is the number of Philosophers. The
move keyword is used to transfer ownership of the Philosopher to the
thread - I think a reference to the silverware vector should be
initialized and moved in the same manner.